Data Access Tokens properties update and endpoint migration to internal API
* Add new property for DAT revocation default value
* Fallback on default value for revoking access tokens on issuance of new token if parameter not provided
* Migrated DAT endpoint to internal API
* Temporary fix to enable DAT config beyond gplus
