+++ security/security-spring/src/main/java/org/cbioportal/security/spring/authentication/googleplus/CustomUserDetailsService.java
@@ -10,8 +10,13 @@ import org.springframework.social.security.SocialUserDetailsService;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
+import org.cbioportal.security.spring.authentication.PortalUserDetails;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
-public class CustomUserDetailsService implements SocialUserDetailsService {
+public class CustomUserDetailsService implements SocialUserDetailsService, UserDetailsService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(CustomUserDetailsService.class);
     
@@ -21,6 +26,8 @@ public class CustomUserDetailsService implements SocialUserDetailsService {
      * @return  The information of the requested user.
      */
 
+    // TO-DO: remove SocialUserDetails?
+
     @Override
     public org.springframework.social.security.SocialUserDetails loadUserByUserId(String userId) {
         Preconditions.checkArgument(!Strings.isNullOrEmpty(userId), "A userid is required");
@@ -28,5 +35,10 @@ public class CustomUserDetailsService implements SocialUserDetailsService {
         
         return new SocialUser(userId,  "unused", new ArrayList<GrantedAuthority>());
         
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return new User(username, "unused", new ArrayList<GrantedAuthority>());
     }
 }
