+++ service/src/main/java/org/cbioportal/service/DataAccessTokenService.java
@@ -37,6 +37,7 @@ import org.cbioportal.model.DataAccessToken;
 
 public interface DataAccessTokenService {
 
+    public DataAccessToken createDataAccessToken(String username);
     public DataAccessToken createDataAccessToken(String username, boolean allowRevocationOfOtherTokens);
     public List<DataAccessToken> getAllDataAccessTokens(String username);
     public DataAccessToken getDataAccessToken(String username);
