+++ service/src/main/java/org/cbioportal/service/exception/DataAccessTokenProhibitedUserException.java
@@ -14,6 +14,22 @@
  */
 
 /*
+ * Copyright 2002-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
  * This file is part of cBioPortal.
  *
  * cBioPortal is free software: you can redistribute it and/or modify
@@ -30,26 +46,16 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package org.cbioportal.weblegacy;
+package org.cbioportal.service.exception;
 
-import org.cbioportal.service.DataAccessTokenService;
-import org.mockito.Mockito;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
+public class DataAccessTokenProhibitedUserException extends RuntimeException {
 
-@Configuration
-@ComponentScan(basePackages = {"org.cbioportal.weblegacy"}, resourcePattern = "**/*DataAccessTokenController.class")
-public class DataAccessTokenControllerConfig {
-
-    @Bean
-    public DataAccessTokenService tokenService() {
-        return Mockito.mock(DataAccessTokenService.class);
+    public DataAccessTokenProhibitedUserException() {
+        super();
     }
-    
-    @Bean
-    public DataAccessTokenController dataAccessTokenController() {
-        return new DataAccessTokenController();
+
+    public DataAccessTokenProhibitedUserException(String message) {
+        super(message);
     }
-    
+
 }
